# core/vm: avoid memory expansion check for trivial ops