This PR moves the dynamic memory check function to only be checked if the op has a dynamic gas function. This is strictly not a guarantee, just how things are, and it allows us to skip two conditionals on the cheap ops. 
One minor semantic change is that `cost` is set earlier, making errors more correct than previously (if erroring on e.g. shallow stack)