@holiman 2021-12-03
Simple loop benchmark
```
[user@work runtime]$ benchstat before after
name                    old time/op    new time/op    delta
SimpleLoop/loop-100M-6     426ms ± 9%     371ms ± 8%  -13.11%  (p=0.000 n=9+10)

name                    old alloc/op   new alloc/op   delta
SimpleLoop/loop-100M-6    2.86kB ±11%    2.91kB ±13%     ~     (p=0.810 n=10+10)

name                    old allocs/op  new allocs/op  delta
SimpleLoop/loop-100M-6      25.2 ± 5%      25.4 ± 6%     ~     (p=1.000 n=10+10)
```