@chfast 2021-12-03
```
name                                  old time/op  new time/op  delta                                                                                                                                                             
micro/memory_grow_mload/nogrow         219µs ± 0%   215µs ± 4%   -1.96%  (p=0.000 n=18+19)                                                                                                                                        
micro/memory_grow_mload/by1            230µs ± 1%   223µs ± 1%   -2.67%  (p=0.000 n=20+20)                                                                                                                                        
micro/memory_grow_mload/by16           292µs ± 1%   285µs ± 1%   -2.39%  (p=0.000 n=19+19)                                                                                                                                        
micro/memory_grow_mload/by32           367µs ± 1%   359µs ± 1%   -2.41%  (p=0.000 n=20+20)                                                                                                                                        
micro/signextend/zero                  390µs ± 1%   324µs ± 1%  -16.93%  (p=0.000 n=20+20)                                                                                                                                        
micro/signextend/one                   406µs ± 1%   337µs ± 1%  -17.10%  (p=0.000 n=20+20)                                                                                                                                        
micro/loop_with_many_jumpdests/empty  1.68µs ± 2%  1.69µs ± 2%     ~     (p=0.583 n=19+20)                                                                                                                                        
micro/jump_around/empty                145µs ± 1%   139µs ± 1%   -4.00%  (p=0.000 n=17+15)                                                                                                                                        
micro/memory_grow_mstore/nogrow        445µs ± 0%   451µs ± 1%   +1.38%  (p=0.000 n=19+19)                                                                                                                                        
micro/memory_grow_mstore/by1           460µs ± 1%   467µs ± 0%   +1.50%  (p=0.000 n=20+19)                                                                                                                                        
micro/memory_grow_mstore/by16          539µs ± 0%   542µs ± 0%   +0.60%  (p=0.000 n=18+20)                                                                                                                                        
micro/memory_grow_mstore/by32          607µs ± 1%   609µs ± 0%   +0.30%  (p=0.006 n=20+20)                                                                                                                                        
micro/JUMPDEST_n0/empty               2.74ms ± 0%  2.58ms ± 0%   -5.96%  (p=0.000 n=19+18)                                                                                                                                        
main/sha1_divs/empty                   196µs ± 1%   179µs ± 1%   -8.52%  (p=0.000 n=20+20)                                                                                                                                        
main/sha1_divs/1351                   3.95ms ± 0%  3.56ms ± 0%   -9.94%  (p=0.000 n=20+16)
main/sha1_divs/2737                   7.71ms ± 0%  6.95ms ± 0%   -9.81%  (p=0.000 n=19+20)
main/sha1_divs/5311                   15.1ms ± 0%  13.6ms ± 0%   -9.75%  (p=0.000 n=18+19)
main/weierstrudel/0                    471µs ± 0%   448µs ± 0%   -4.72%  (p=0.000 n=18+20)
main/weierstrudel/1                    980µs ± 1%   918µs ± 0%   -6.31%  (p=0.000 n=20+18)
main/weierstrudel/3                   1.54ms ± 1%  1.45ms ± 1%   -5.90%  (p=0.000 n=20+19)
main/weierstrudel/9                   3.18ms ± 0%  3.02ms ± 1%   -5.24%  (p=0.000 n=20+18)
main/weierstrudel/14                  4.55ms ± 1%  4.32ms ± 1%   -5.13%  (p=0.000 n=18+20)
main/sha1_shifts/empty                 151µs ± 0%   132µs ± 1%  -12.63%  (p=0.000 n=20+20)
main/sha1_shifts/1351                 3.09ms ± 0%  2.68ms ± 1%  -13.24%  (p=0.000 n=14+20)
main/sha1_shifts/2737                 6.02ms ± 0%  5.22ms ± 0%  -13.34%  (p=0.000 n=17+20)
main/sha1_shifts/5311                 11.8ms ± 1%  10.2ms ± 1%  -13.34%  (p=0.000 n=19+20)
main/blake2b_huff/empty                100µs ± 0%    98µs ± 0%   -1.48%  (p=0.000 n=18+19)
main/blake2b_huff/2805nulls           1.90ms ± 1%  1.88ms ± 1%   -1.08%  (p=0.000 n=20+19)
main/blake2b_huff/5610nulls           3.70ms ± 0%  3.65ms ± 1%   -1.44%  (p=0.000 n=19+20)
main/blake2b_huff/8415nulls           5.43ms ± 1%  5.34ms ± 0%   -1.65%  (p=0.000 n=20+17)
main/blake2b_shifts/2805nulls         14.2ms ± 2%  12.9ms ± 1%   -9.21%  (p=0.000 n=20+17)
main/blake2b_shifts/5610nulls         28.1ms ± 0%  25.7ms ± 1%   -8.44%  (p=0.000 n=18+20)
main/blake2b_shifts/8415nulls         41.9ms ± 0%  38.2ms ± 1%   -8.73%  (p=0.000 n=16+17)
[Geo mean]                            1.14ms       1.07ms        -6.19%
```